requirements.txt
requests>=2.31.0
pandas>=2.0.0
dbt-bigquery>=1.7.0
google-cloud-bigquery>=3.10.0
google-auth>=2.0.0
apache-airflow==2.9.3
apache-airflow-providers-google>=10.0.0
sqlalchemy>=1.4.0,<2.0.0



















